refactor: Relocate playlist commands and remove aliases from slash commands
