+++ src/scripts/commands/registerGlobalSlashCommands.ts
@@ -15,16 +15,8 @@ void (async () => {
     const paths = await subfilePathsOf("commands");
     Logging.info("Start registering commands...");
     for (const path of paths) {
-        console.log(path);
         const command = await importDefault<Command>(path);
         if (!command) break;
-        if (command.aliases) {
-            for (const alias of command.aliases) {
-                const aliasCommand = Object.assign({}, command);
-                aliasCommand.name = alias;
-                commands.push(aliasCommand);
-            }
-        }
         commands.push(command);
     }
