--- src/typings/commands/music/ListPlaylistOptions.ts
@@ -1,5 +1,0 @@
-import type { NoOptions } from "../../structures/commands/SmoothieCommand.js";
-
-type ListPlaylistOptions = NoOptions;
-
-export default ListPlaylistOptions;
