+++ src/typings/commands/music/playlist/ListPlaylistOptions.ts
@@ -1,0 +1,5 @@
+import type { NoOptions } from "../../../structures/commands/SmoothieCommand.js";
+
+type ListPlaylistOptions = NoOptions;
+
+export default ListPlaylistOptions;
